Safari-609.1.11.1/LayoutTests/http/tests/workers/location-readonly.html

Test passes if "127.0.0.1" is returned by the worker.
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}
w = new Worker('worker-location.js');
w.onmessage = function(e){
    console.log(e.data);
    testRunner.notifyDone();
}
</script>
