Safari-609.1.11.1/LayoutTests/http/tests/workers/location-readonly-expected.txt

CONSOLE MESSAGE: line 9: 127.0.0.1
Test passes if "127.0.0.1" is returned by the worker.
